+++ DurableBetterProspecting/Core/SampleShape.cs
@@ -3,6 +3,6 @@ namespace DurableBetterProspecting.Core;
 internal enum SampleShape
 {
     Vanilla,
-    Cube,
-    Cuboid
+    Sphere,
+    Cylinder
 }